# http/swagger_gen/api/swagger.yaml
definitions:
  echo_example.Product:
    properties:
      content:
        items:
          type: object
        type: array
      id:
        type: integer
      interest:
        type: number
      is_active:
        type: boolean
      title:
        type: string
      total:
        type: integer
    type: object
info:
  contact: {}
  title: APIExample
  version: "1.0"
paths:
  /file:
    get:
      description: Send a file
      operationId: DownloadFile
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          headers:
            Content-Disposition:
              type: string
            Content-Type:
              type: string
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send a file
      tags:
      - files
  /products:
    get:
      consumes:
      - application/json
      description: Get a list of products
      operationId: ListProducts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/echo_example.Product'
            type: array
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product
      operationId: CreateProduct
      parameters:
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/echo_example.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/echo_example.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      operationId: DeleteProduct
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a product by ID
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get a product by ID
      operationId: GetProduct
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo_example.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update a product by ID
      operationId: UpdateProduct
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/echo_example.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo_example.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a product by ID
      tags:
      - products
swagger: "2.0"